refactor: 네이밍 리펙토링 및 컨트롤러 응답 변경 (#98)
